Add mandatory :key attribute to README template
I previous opened issue #12, but turns out it's a simple oversight. I had followed the example template in the README and being new to Vue, did not realize that `:key` was mandatory. I'm adding this to the README so that others that are new to Vue won't spend hours trying to debug such a simple issue...